Change `linear` to use Closes: instead of force-pushing to master
One of my projects (projectatomic/rpm-ostree) has a protected master
branch.  I was rather horrified to realize that homu force-pushes to
master by default.  This ended up leaving the "temporary" commit on
master.  Admittedly, I understand this is the only way to make github
think the PR was merged (as opposed to just closed).

Force pushing to master is IMO a *terrible* idea becuase it means
clients doing pulls might see the transient commit, etc.

What we really want is a github API to say that we did a merge,
but until that exists, I think using the `Closes` tag is good enough.
We also tweak the PR title to include `[merged]` which helps a lot.

This is a rework of a prior commit which made this behavior an option,
but I think force pushing is a bad enough idea that we should just
replace the old code.  If anyone complains, we can discuss options.