Try builds should be able to use a different branch than auto
There is a race condition on all of our CI systems when there is a build queue
and multiple builds (e.g., try and auto) pushing to the same brach. If one
build is pushed, a CI is queued but not started, and then another build is
pushed to the same branch, when the queued CI job begins executing, it will
fail to find the git commit hash since it is no longer in the branch.