Support overriding git username+email in config
When one creates an automated github account for Homu, it may be
desirable to have a private email associated with the github account.

Currently it seems e.g. the Servo project's bors-servo does not have a
public email instance, but we're relying on the app ID being granted
access to the `user` scope to see it, so Homu ends up writing
lbergstrom's email into the commit messages, even though it's not
public.

I suspect most people will prefer what this patch offers, which allows
directly setting the git user/email (rather than retrieving it from
github), so one can easily use e.g. a public mailing list for the
commits.  Which would be a more useful place to send mail generally.
Anyone who wants to contact the bot owner should be able to figure out
how to do that.