main: Support a -c/--config argument
I've been previously running Homu in Docker in a somewhat
hacky way, and I'm trying to adapt it into a proper OpenShift/Kubernetes
setup where the cfg file is a provisioned secret (read-only).

Currently the way things are hardcoded to have the config and the
database (which is really conceptually a cache), colocated makes
implementing this very hard.  So add an option.