main: Do authorization per-command
One really nice thing this fixes is that before now, if a non-reviewer tried
`r+`, Homu would silently ignore them, leading to a mental state somewhere
between confusion and rage. Now we reply with an unauthorized message,
so at least they won't be confused. There might still be rage though, but
nothing we can do about that here!

Further, this sets the stage down the line for having new commands with
different types of authorization. For example, maybe we automatically allow the
PR creator to do `retry`.

But what I'm really going for in the next patch is to add an option to back this
with a dynamic query to see if the commenter is a collaborator. Doing that would
have been expensive if we did it every time we parsed a comment, before we even
knew whether or not it was a command.