main: Support auth_collaborators
There are two reasons. First, github already has a model for who can commit to a
repo, and it's a lot easier to incrementally adopt Homu to be able to reuse it.
It makes it easier to fall back in emergency situations if the Homu instance
goes down, etc.

Second, having to restart Homu to add a reviewer is painful.