When sorting, lump rollup=never and rollup=maybe
When determining the sort order for pull requests, always sort
`rollup=always` to the bottom of the list, and ignore any other rollup
status.

Previously, a PR with `priority=0`, `rollup=never` would be in the queue
above a PR with `priority=1`, `rollup=maybe`. This commit makes it so
that PRs with `rollup=never` is **not** given higher weight than
`rollup-maybe`, so the `priority=1` pull will rank higher.

Fixes #27