Restyle the queue page
Change the CSS styles for the queue page to bring a more clean look.

To prevent disruption, do not change the flow of the page, the locations
of any of the columns, or any of the key colors.

Attempt to make it clear to newcomers that the "pending" state is the
one where jobs are currently running by adding a subtle animation to the
status indicator.

Attempt to highlight priority jobs by displaying non-zero priorities
differently from zero priorities.