Fix build result pages for pending jobs
We have a link on the queue page for success, pending, and failed jobs.
Links to pending jobs seem to never work. This seems to be because we
return HTTP status code 204, which means that no content is available.

Update the result page to always attempt to show a table instead of
returning a 204 when we know about the pull request, even if we have no
build statuses.

Update the result page to return a 404 instead of a 204 when we have no
information on the pull request. This gives some information to the end
user, instead of appearing to do nothing.