Replace db_query with LockingDatabase
Remove the free function `db_query` in favor of a `LockingDatabase`
class that wraps a database connection.

This is done so that, in order for a class to use the database, it only
needs its `db` instance, instead of needing both a `db` instance and a
`db_query` function.

This allows us to break out classes into separate files.