GitHub as event stream proof-of-concept
We currently have two sources-of-truth when syncing on startup: the
database and issue comments.

This proof of concept tries to bring us down to a single source-of-truth
by effectively elimating the local database (at least for pull request
state; other aspects like `treeclosed` are not addressed).

In order to do this, we use GitHub's "timeline"
([v3](https://developer.github.com/v3/issues/timeline/),
[v4](https://developer.github.com/v4/object/pullrequest/#timelineitems))
feature to get a fuller view of what has happened to a pull request so
we can reify state in a more complete manner than using issue comments
alone. For example, the timeline API interleaves issue comments with
things like `AssignedEvent`, `HeadRefForcePushedEvent`, and
`RenamedTitleEvent` in the order they happened so we can reliably build
the full state from this historical record.

The general concept is that in both `synchronizing` and `realtime` mode,
we have the pull request object respond to the event to update the
state. Side effects like comments and label updates are returned, and
the code can use or drop those depending on which mode Homu in.

    PR state --------\
                     |     +---------------+
                     \---> |               | -----> Modified PR state
                           | process_event |
                     /---> |               | --\
                     |     +---------------+   +--> Realtime comments
    GitHub Event ----/                         +--> Label updates
                                               \--> Database update

The good thing is that we can model these timeline items and GitHub's
event webhooks (which we use when processing in `realtime` mode) in much
the same way, so that we can unify them in a way that makes it possible
to effectively test the effect of events on a pull request.

This particular pull request does not attempt to integrate these changes
into the codebase yet, because I wanted to get feedback that this is
indeed a viable path to head down.

Instead, this pull request (which should not be merged) provides a new
executable that demonstrates this ability on existing pull requests in a
read-only mode. The output is a series of relevant timeline items, and
the state of the pull request as it changes.

    pip install -e homu
    homu-sync-test --repo rust-lang/rust 61419