Test authorization code
Pull code that creates a comment on a pull request out of the
authorization check, and add tests around the authorization checks.

After pulling out the code that creates a comment, we still need to know
what the text of the comment will be, so change from a function that
returns a True on success and a False on failure to a function that
returns a True on success and raises an Exception (with the failure
comment) on failure.