Parse homu state out of comments
For some critical comments, Homu adds extra information about its state
in the form of a JSON blob to the comment that isn't visible to the user
but is visible in the source for the comment. For example, Homu may
leave a comment like the following, where the JSON blob is not visible
because of the `<!-- .. -->` markdown/html comments.

    :hourglass: Trying commit abcdef with merge 012345...

    <!-- homu: {"type":"TryBuildStarted","head_sha":"abcdef","merge_sha":"012345"} -->

This change parses this extra information out of the comments and makes
it available to the initial synchronization algorithm.