Stop parsing when we reach an unknown word
When parsing a @bors command, stop parsing when we reach any unknown
word. This enables commands like

    @bors retry (yielding priority to the rollup)

to include a description of why we retried (which gets put in the retry
log) without also setting `rollup`.

The only tricky bit to this is a command like

    @bors r=person 0123456789abcdef

where the commit sha is part of the `r=` command, and we need to parse
it that way.