Fix repository type error
Fix a type error where we (potentially) mistakenly assign a
`github3.repos.repo.Repository` object to a place where we expect a
`Repository` (as defined in main.py).

It appears this was intended in 2067c37d but missed.

Also fix a typo in a variable name that doesn't seem to have caused
problems yet, but let's be safe.