Add test to ensure @bors only responds to itself
Test that we only respond when a command is issued to `@bors`, and not
to `@bors-servo` or any other `@bors{something}` bot.

This came up because a comment on `rust-lang/rust` referenced
`@bors-servo` instead of `@bors` and was still approved successfully.
The command-parsing changes apparently already fixed this, but add a
test to prevent regressions.