Include start/end times in metadata comments
Timing information could be gleaned from comment times, but make it
explicit in the metadata. This means we can get the true times, even if
GitHub has delays.

This also gives us the opportunity to have more control in the future.
For example, the [check_run event][check_run] includes `started_at` and
`completed_at` information that is independent from when GitHub sent us
the webhook.

[check_run]: https://developer.github.com/v3/activity/events/types/#checkrunevent-api-payload