Calculate `status`, `try_` and others from history
Previously, we discretely set `status`, `try`, `build_state`, and
`try_state` on each event. With the addition of the run histories, we
can now glean all of this information from those histories instead of
tracking them independently.

The results appear to all be identical on the current Homu queue except
for one edge case: a pull request was tried, approved, then unapproved.

- Using the previous method, the status would be '', because approval
  changes the status from 'success (try)' to '', and unapproval doesn't
  change the status.

- Using the current method, the status would be 'success (try)', because
  a successful try has occured for the relevant commit hash and the PR
  isn't approved.