Shift success rows in queue to bottom
Generally these are try builds which have concluded; ideally we wouldn't show
them at all (there's really no reason for them to be in the queue) but that's a
harder change. On the rust-lang/rust queue these typically take up the first 1.5
pages of queue -- which is wasteful, as there's not really interesting
information to be gleaned from them.

This moves that list to the bottom of the page.